Breakout `groupRectangles()`
